[Flow262][skip-ci] Fix colors.js errors
Summary:
https://linear.app/comm/issue/ENG-10205/update-flow-to-02691

`type Colors = $Exact<typeof light>;` now resolves to an object with literals like:
```
 {
  blockQuoteBackground: '#...',
...
```
So we can't use it to type `dark` theme object.
Also there are not only strings but also numbers in the `Colors` type.
So sadly I had to create the `Colors` type manually.

Depends on D15018

Test Plan: Flow

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D15019