[native] Fix restore API call order
Summary:
Noticed that in D14752 the order of the calls was correct in one file but wrong in the other. This resulted in
```
LOG  Error when restoring User Data: restoredConnectionManager is not set
```

Test Plan: The error no longer happens and after restoring (primary login) I'm able to see restored content (e.g. drafts)

Reviewers: kamil!

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D14794