[web] implement encryption of restored database
Summary:
[ENG-10786](https://linear.app/comm/issue/ENG-10786/database-downloaded-from-the-backup-is-not-encrypted).

Refactored a bit to run two promises simultaneously.

We're using the same key as for the main database, because there is a bug on Safari which requires generating keys on the main thread ([ENG-3872](https://linear.app/comm/issue/ENG-3872/browser-worker-communication-sometimes-fails-on-safari)), and I think using the same key here should be fine, as restored database is temporary and we persist it only to restore in case of app being killed in the middle of restore.

Depends on D14804

Test Plan: Test if persistence still works, test restore on web

Reviewers: bartek, tomek

Reviewed By: bartek

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D14805