[lib] gate `SyncStoreVersionHandler` on rehydration
Summary:
[ENG-10816](https://linear.app/comm/issue/ENG-10816/update-runmigrations-to-the-new-approach).

After looking deeper into removal og updating synced metadata entries as part of migration ([link](https://github.com/CommE2E/comm/blob/4487e6d0abbb0347543fb42f1afc142eccacbb52/lib/utils/migration-utils.js#L307-L313)), I realized we shouldn't remove it code and there is one problem with `SyncStoreVersionHandler`.

With current implementation, there is a slight chance of inconsitency in synced metadata entry. In the past, handler was always setting the value to the `stateVersion` in codebase, regardless of whether migration succeded or not, or if app was killed in the middle of migrations.

With this fix, we first run migraitons, each migrations bumps state version in the database to its version and after this handler is executed.

The change of this is very low because handler is gated on `ReduxPersistGate` on native, and in case of migration failure we call `handleReduxMigrationFailure` but worth doing this in a safe way.

Depends on D14807

Test Plan: Add migration and make sure that handlers runs after the migration.

Reviewers: tomek, bartek

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D14808