[lib] gate `SyncStoreVersionHandler` on rehydration
Summary:
[ENG-10816](https://linear.app/comm/issue/ENG-10816/update-runmigrations-to-the-new-approach).

After looking deeper into removal og updating synced metadata entries as part of migration ([link](https://github.com/CommE2E/comm/blob/4487e6d0abbb0347543fb42f1afc142eccacbb52/lib/utils/migration-utils.js#L307-L313)), I realized we shouldn't remove this code and there is one problem with `SyncStoreVersionHandler`.

With the current implementation, there is a slight chance of inconsistency in synced metadata entry. In the past, the handler was always setting the value to the `storeVersion` in the codebase, regardless of whether migration succeeded or not, or if the app was killed in the middle of migrations.

With this fix, we first run migrations, each migration bumps the state version in the database to its version, and after this handler is executed.

The chance of the issue I described is very low because the handler is gated on `ReduxPersistGate` on native, and in case of migration failure, we call `handleReduxMigrationFailure`, but worth doing this in a safe way. The code around might change in the future, and this gives us assurance that we have the correct `storeVersion` in the backup database.

I am also wondering if code in [/native_rust_library/build.rs](https://github.com/CommE2E/comm/blob/e4174e16d28d3a385c2f05f326c0e37017678d1e/native/native_rust_library/build.rs#L130-L161) has the same issue, but not sure what the impact of `storeVersion` in Rust

Depends on D14807

Test Plan: Add migration and make sure that handler runs after the migration.

Reviewers: tomek, bartek

Reviewed By: bartek

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D14808