[lib] Fetch keys only when a session is missing
Summary:
If a session was missing for any device, we were asking for the keys for all the devices.

https://linear.app/comm/issue/ENG-10753/session-recovery-sometimes-fails

Test Plan: Create a session between device A1 of the user A and the device B1 of the user B. Then add a new secondary device B2 and send a message from A to B. Make sure using some console logs that only keys for B2 were fetched and that only this session was created.

Reviewers: kamil, bartek, ashoat

Reviewed By: ashoat

Differential Revision: https://phab.comm.dev/D14816