[Flow262][skip-ci] Fix types in store migration
Summary:
https://linear.app/comm/issue/ENG-10205/update-flow-to-02691

I know we should avoid `as any` but here we're dealing with migrations from the old state to the new state and the old state might not match the current state type.
Let me know if there is a better solution for this.

Depends on D15047

Test Plan:
Flow

Reviewers:

Subscribers: