[lib][web][native] update `ReplaceMessageStoreThreadsOperation` to decide if data should be included backup
Summary:
[ENG-10847](https://linear.app/comm/issue/ENG-10847/update-replacemessagestorethreadsoperation-to-decide-if-data-should-be).

I decided to go for the approach when we split one operation into two, where in one there are only backed up threads and in the other operations only threads that are not.

Depends on D14825

Test Plan: It's really hard to test everything here; for now, only Flow. In the final diff of the stack, I am going to do some more testing by testing core functionalities of threads and messages (including local and media messages).

Reviewers: tomek, bartek

Reviewed By: tomek

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D14826