[SQLite][web] implement tests for `getRelatedMessages()`
Summary:
[ENG-10464](https://linear.app/comm/issue/ENG-10464/biggest-risk-handle-target-message-in-messages-table#comment-3e544d1d).

In the next diff, I am going to change the way `target_message` is handled. Looking at the code and in history, when `target_message` was introduced, it was mainly for `getRelatedMessages` query (stack ending with D12737).

To be able to test it somehow, I decided to first implement detailed tests on how `target_message` is utilized - I used Claude code for this, and it did a surprisingly good job.

Test Plan: Run Tests

Reviewers: tomek, bartek

Reviewed By: tomek

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D14835