[native] Add primary device check to DatabaseManager
Summary:
Address [[ https://linear.app/comm/issue/ENG-8227/implement-jsi-to-check-if-we-are-running-on-primary-device | ENG-8227 ]].
Created C++ method to determine if we're running DB operations on a primary device.

Depends on D14833

Test Plan:
Used Xcode debugger to verify that the function works as expected. Tested primary and secondary device.
Made sure logs are not captured on secondary device

Reviewers: kamil

Reviewed By: kamil

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D14834