[SQLite] update deleting queries to support backup tables
Summary:
[ENG-10851](https://linear.app/comm/issue/ENG-10851/update-functions-that-can-be-executed-regardless-of-the-fact-whether).

Those methods are executed as part of operations that are wrapped in a transaction, so it is safe to implement them that way.

It is possible to implement queries as one string (queries separated by `;`) to reduce code a bit, but this will require huge changes in `EntityQueryHelpers`, and I am not sure if it is worth it. From a performance perspective, it shouldn't be much worse.

Depends on D14841

Test Plan: Tests in D14848

Reviewers: tomek, bartek

Reviewed By: tomek, bartek

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D14844