[SQLite] update replace queries to support backup tables
Summary:
[ENG-10852](https://linear.app/comm/issue/ENG-10852/update-functions-that-has-to-be-executed-on-specific-tables).

For now, using `false` in `backupItem` flag everywhere. It is going to change when working on [ENG-10850](https://linear.app/comm/issue/ENG-10850/updated-classes-in-datastores-to-handle-param-is-data-should-be). This is going to be the `dataIsBackedUp` prop from D14820 passed to SQLite.

`updateSQLiteUnreadStatus` and `storeMessageInfos` should be handled separately. I created a task for it: [ENG-10952](https://linear.app/comm/issue/ENG-10952/handle-native-api-that-modifies-database-directly)

Depends on D14846

Test Plan: Tests in D14848

Reviewers: tomek, bartek

Reviewed By: tomek

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D14847