[SQLite] update `getRelatedMessages` to support regular and backup tables
Summary:
[ENG-10858](https://linear.app/comm/issue/ENG-10858/update-getrelatedmessages-to-support-multiple-tables).

Depends on D14855

Test Plan: Experiments with `EXPLAIN QUERY PLAN` to check if indexes are used and performance is not worse. I decide to not update test because it'll be redundant. Instead, what I did is go to `get-related-messages.test.js` (added in D14835) and updated temporary flag (added in D14847) to add messages to the backup tables and run tests again.

Reviewers: tomek, bartek

Reviewed By: tomek

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D14856