[services] Exclude .idea project from building
Summary:
When using JetBrains tools, `.idea` directory is created to keep some metadata about a project. Our Cargo setup was treating this as a new project and failed building because there was no `Cargo.toml` inside. The solution is to exclude this directory from building.

Test Plan: Make sure that the directory is present and build the project.

Reviewers: kamil, bartek