[lib] set `dataIsBackedUp` to be false in `keyserverThreadProtocol`
Summary:
This is an oversight from D14820, it was hard to discover it because I wasn't able to test the actual impact on backup then. This is still ignored on the database layer (still in progress), so there is no regression on master.

Depends on D14820

Test Plan: N/A

Reviewers: tomek, bartek

Reviewed By: tomek

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D14853

(cherry picked from commit 6e66b7e3a802776e2fb72a8dbcefaed17c12013f)