[native] implement `threadIDMatchesProtocol` in C++
Summary:
It is going to be used in [ENG-10859](https://linear.app/comm/issue/ENG-10859/update-searchmessages-to-support-multiple-tables) and in [ENG-10952](https://linear.app/comm/issue/ENG-10952/handle-native-api-that-modifies-database-directly).

There is also some context why this is required in [this comment](https://linear.app/comm/issue/ENG-9787/update-functions-that-read-from-the-db-to-be-aware-of-thread-thickness#comment-6a628ea0).

Depends on D14856

Test Plan: Tested in next diff

Reviewers: tomek, bartek

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D14857