[native] Add callback on compaction creation
Summary:
This is for compaction part of [[ https://linear.app/comm/issue/ENG-7012/restarting-backup-restorationupload | ENG-7012 ]]. We might want to resume existing compaction when it was created but hasn't been uploaded.
Original compaction promise resumes after successful upload, so added a callback that is triggered after compaction is created but before it's uploaded.
To avoid significant changes to RustPromiseManager, on C++ side, I had to create a dummy Promise that calls the callback when resolved.

Depends on D14886

Test Plan: Triggered backup from backup menu. Confirmed that console log from the callback has the same backupID that is later resolved when the upload promise fulfills.

Reviewers: kamil

Reviewed By: kamil

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D14887