[native] Return backupID from restoreUser() Identity call
Summary:

When restoring user, we create a new UserKeys backup and generate new backupID on the native side, but we never return it to the JS caller. Updated the types to return it.
This is useful in the next diff where we store it.

Test Plan:

Tested together with the next diff.

Reviewers: kamil

Subscribers: