[Flow262][skip-ci] Fix thread infos validators
Summary:
https://linear.app/comm/issue/ENG-10205/update-flow-to-02691

Deleted ThinRawThreadInfoWithoutSpecialRole because it seems to be no longer needed and it fixes a lot of issues with thread validators. It was introduced in this diff: https://phab.comm.dev/D11496 and from what I've read it was introduced because there was a change in RolesInfo. Old clients expected RoleInfo with isDefault and new clients expected RoleInfo with specialRole. That's why the keyserver sends role infos with isDefault for old clients: https://github.com/CommE2E/comm/blob/master/lib/shared/thread-utils.js#L822. shouldIncludeSpecialRoleFieldInRoles is true from version:
```
native: 336,
web: 79,
```
and we're no longer supporting those versions. Also we don't use those types and validators anywhere anymore, so it should be ok to delete them.

As for `rawThreadInfos` in `StateChanges` I changed validator to conform to types and it's broader now so there shouldn't be any issues with types.

Depends on D15065

Test Plan: Flow

Reviewers: tomek

Reviewed By: tomek

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D15066