[tunnelbroker] Increase a heartbeat timeout
Summary:
It may happen that a client is occupied, and it takes it more time to respond to a heartbeat. This may happen e.g., when a longer job is scheduled on the event loop.

Bumping this more might be dangerous because of the client timeout https://github.com/CommE2E/comm/blob/6c9b95cac16ea755b04b91b9bcd0537fddfe33b3/lib/shared/timeouts.js#L42.

https://linear.app/comm/issue/ENG-10975/increase-tunnelbroker-heartbeat-timeout

Test Plan: Added some console logs to verify that the heartbeats are sent less frequently.

Reviewers: kamil, bartek

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D14904