[lib][web][native] remove holders asserting and blacklist holder store
Summary:
I added this in D14896 to follow some patterns from the past, but it turned out to add complexity in the case of backup. When restoring, we dispatch exactly the same action as on app start ([see code here](https://github.com/CommE2E/comm/blob/2533c1ca033eb87a24ecd78ef25f546b5a2dcac3/lib/backup/use-user-data-restore.js#L122)), which means we get exactly the same alert as in case of inconsistency each time, which is confusing.

Historically, asserting was used to make sure reducer refactor to ops pattern works fine, and there are some ways of handling this in code, but Holders reducer is easy, I tested all paths and we also have tests so I think we can just remove it to avoid adding and maintingn code which value is minor.

Depends on D14902

Test Plan: Send multimedia and log holders, then kill and open app and verify the holder is there.

Reviewers: tomek, bartek

Reviewed By: bartek

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D14905