[web] Display PushNotif modal after restoration has finished
Summary:

Very similiar to D14917 and D14918. Without this change, the push notif modal was shown in the middle of restoration progress UI (to be added later).
This has no effect when `fullBackupSupport` is false (dataReady is always true).

Depends on D14918

Test Plan:

With fullBackupSupport, the push notif modal is shown after `useUserDataRestore()` hook finishes execution.

Reviewers: kamil

Subscribers: