[lib] avoid overriding existing messages on `setClientDBStoreActionType`
Summary:
[ENG-10989](https://linear.app/comm/issue/ENG-10989/cannot-read-properties-of-undefined-reading-time-when-restoring-backup).

Unfortunately, I am no longer able to reproduce the original issue from the task description (even using the same account and backup), but because of my observations when creating the task, I was able to find the issue.

Steps to reproduce:
1. Send more than 20 messages in thin thread.
2. Restore backup and add `sleep` when reading content from User Data.
3. There is no UI yet, so before dispatching data from backup, you can see all messages in this thin thread.
4. After data from the backup is dispatched, there are only 20 messages.

This is because `getClientDBStore`, for thin threads, returns only 20 messages and `setClientDBStoreActionType` is overriding all previous messages.

This diff fixes this, for `messageStore.threads` this is handled correctly in [here](https://github.com/CommE2E/comm/blob/caa5fc02d9aa796ee64e573ae1026f0a62fba3a7/lib/reducers/message-reducer.js#L1734-L1736).

This issue is not occurring for other reducers, because for other data types we read everything as part of `getClientDBStore`, only for messages we read a subset.

Test Plan: Issue is gone

Reviewers: bartek, tomek

Reviewed By: bartek

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D14929