[lib] implement `BackupIsNewerError` class
Summary:
[ENG-10812](https://linear.app/comm/issue/ENG-10812/store-the-information-that-backup-is-not-restored-because-app-is-out).

This is a more convenient way of handling this. Additionally, we can store it in `user_data_restore_failed` payload.

This is going to be handy as part of [ENG-10817](https://linear.app/comm/issue/ENG-10817/prevent-users-from-using-the-app-before-user-data-is-restored), where we want to keep a modal that the app needs to be updated and preventing users from using the app.

Depends on D14936

Test Plan:
1. Bump store version
2. Upload backup
3. Restore store version to lower
4. Build app & try to restore, alert should be visible. Additionally check state if proper error is stored

Reviewers: bartek

Reviewed By: bartek

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D14937