[lib] update `broadcastEphemeralMessage` to return result
Summary:
In next diff, we want to use it to send backup keys to secondary devices. We want to know when it succeeds and when it fails, so it can be restarted and we can be sure that the message was delivered.

Depends on D14947

Test Plan: Tested in next diff

Reviewers: bartek

Reviewed By: bartek

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D14948