[lib] implement handling backup keys from primary
Summary:
[ENG-10632](https://linear.app/comm/issue/ENG-10632/biggest-risk-implementing-a-handler-in-react-that-will-perform-a-one).

This change should cause showing appropriate UI to show success or the reason for failure (cc. @bartek).

In case of a secondary device getting this message after a new compaction was uploaded in the meantime, it will work because, for now, in the initial phase, we keep all the backups, so it will just restore from the old one.

After [Cleaning up old backups](https://linear.app/comm/project/cleaning-up-old-backups-34555cb72d32/overview), it will still show UI with error and instructions for the user, same as now when it fails during e.g. secondary device login.

Depends on D14949

Test Plan:
1. Restore on primary device
2. Add some secondary devices
3. Flip `fullBackupSupport`
4. User Data uploaded without any additional changes
5. List of secondary devices is in the store
6. Message with keys is send to all secondary devices, only once. In case of missing session or encryption issue it works too.
7. All data that was missing on secondary, like messages is already there

Reviewers: bartek

Reviewed By: bartek

Subscribers: ashoat, tomek, bartek

Differential Revision: https://phab.comm.dev/D14950