[lib] Inform other own devices about the connection update
Summary:
After connecting or disconnecting, send a message to all the other devices about new connection status.

https://linear.app/comm/issue/ENG-10903/inform-all-the-users-devices-about-the-successful-connection

Depends on D14984

Test Plan: Log in on two devices - at least one physical (so that the Farcaster app can be used). Open the Farcaster section in the profile on the simulator and notice how it changes when performing operations on the physical device. On the physical device connect Farcaster, connect DCs, and disconnect - the result should be reflected on the simulator.

Reviewers: kamil, bartek

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D15093