[lib][native][web] Add a way to override Identity AuthMetadata
Summary:
A way to workaround a situation, where CSAT is not yet available in Identity client because the context component / hooks haven't re-rendered yet after credentials update. This is especially an issue in long async functions, like the whole restore protocol, where auth metadata is set in the middle of the process.
The solution is not the nicest but it works well in such situations - "metadata override" can be provided manually, to avoid `getAuthMetadata()` throwing "Missing auth metadata".

Depends on D14934

Test Plan: Restore flow - QR auth on both platforms + primary restore on native. The 'update peers' step now always succeeds

Reviewers: kamil

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D15146