[native] improve handling `BackupDataP2PMessage` on primary
Summary:
[ENG-11025](https://linear.app/comm/issue/ENG-11025/improve-handling-backupdatap2pmessage-on-primary).

This is safer, because in case of session failure this message could be encrypted again and resent.

There is an alternative way, to completely remove handler and implement it fully in reducer (reducers can remove `outboundP2PMessages`), but fetching `backupData` there and doing all the other checks will make code a lot more complicated.

Test Plan: Did same test plan as in D14949

Reviewers: tomek, bartek

Reviewed By: bartek

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D15165