[tunnelbroker] implement Farcaster client
Summary: [ENG-10869](https://linear.app/comm/issue/ENG-10869/farcaster-communication-layer).

Initial version of client, later in the stack requests are built on clients, but after changing it to a different approach where this is handled by Tunnelbroker, it could be reused, only `payload` can change to something with stronger types.

Additionally, here token is read on each request, which makes it slow but resistant to race condition and token changes. After working on what @tomek suggested in [here](https://linear.app/comm/issue/ENG-11060/consider-tunnelbroker-role-in-farcaster-dc-integration#comment-56be1678)this could be improved and cached

Test Plan: Tested later in the stack with client code

Reviewers: bartek, tomek