[Flow262][skip-ci] Silence for..in errors
Summary:
https://linear.app/comm/issue/ENG-10205/update-flow-to-02691

Flow has broken for..in loops: https://github.com/facebook/flow/issues/1736
For the loop like:
```
for(const key in obj)
```
`key` is always resolved as `string` and not literals even when keys of `obj` are literals. And now Flow is more strict about string vs literals.
So I've tried to workaround the problem by using `Object.keys` and `foreach` where possible and silencing the error when necessary.

Depends on D15021

Test Plan: Flow

Reviewers: tomek

Reviewed By: tomek

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D15022