[lib] improve handling backup keys from primary when user proceeded without restoring
Summary:
[ENG-11091](https://linear.app/comm/issue/ENG-11091/secondary-device-is-missing-messages-in-local-dms).

This fixes the case when @ashoat logged in on the web with a full backup released before the primary uploaded backup. This worked as expected; there was a prompt with an error, and the user can decide whether to proceed with backup or ignore this and contact Comm.

Later, when the updates to primary keys were sent to secondary, they were ignored because proceeding set app was in the "restored" state. However, we store the information in the store ( `forced` flag) to distinguish regular restore vs restore forced by user to log in and ignore backup issues. Based on that flag we can find out when restore wasn't concluded and avoid ignoring backup keys.

Test Plan:
1. Update `fullBackupSupport` to return `true` on web and `false` on native
2. Create some DMs messages
3. Add secondary device `web`, there is alert shown that backup restoration failed (it is not there, which is correct). DMs messages are missing
4. Update primary
5. On secondary there is UI about restoring, all DMs messages are there

Reviewers: bartek

Subscribers: tomek, ashoat

Differential Revision: https://phab.comm.dev/D15193