[lib] Fix restore screen flashing during logout
Summary:
Fixes [[ https://linear.app/comm/issue/ENG-11086/restoring-your-data-screen-flashes-on-web-after-logout | ENG-11086 ]].
On web, the display logic is based on this value, so when `loggedIn == false && restoreBackupState != 'no_backup'`, progress screen is displayed.

We should set restore state to `no_backup` as soon as logout starts. This both fixes the display issue and is more correct in terms of app/backup lifecycle logic.

Test Plan:
- Restored a primary device.
- Logged in a secondary web device. UI displays correctly, no changes in previous behavior.
- Logged out on the web device.
  - Before: Restore screen is displayed between `logOutActionTypes.started` and `logOutActionTypes.success`
  - After: QR code screen is displayed immediately, [a spinner is displayed](https://github.com/CommE2E/comm/blob/1a86cbe9dfa11b8971ab161302d476267e2623f5/web/account/log-in-form.react.js#L163) in place of the QR code until the fresh DB is ready to log in again
- Logged out primary device and restored again. This works as before, nothing changed here.

Reviewers: kamil

Reviewed By: kamil

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D15195