[lib] Don't treat initial missing olm session as failure
Summary:

This is to reduce log spam, since this situation is usually normal (we optimistically), it just happens more often after D15140 which skips session creation for "dead" (non-existent) devices.
When a device is missing, device list is refetched and the messages are later encrypted with the updated device keys.

Test Plan: Checked that `session is missing` is now Olm log instead of Olm Error in debug logs

Reviewers: kamil, tomek

Subscribers: