[comm-lib] add table for Farcaster token
Summary: [ENG-11096](https://linear.app/comm/issue/ENG-11096/add-new-ddb-table-for-farcaster-dcs-tokens)

This is going to be used to distribute tokens between Tunnelbroker instances in order to open socket connections.

In this solution, Tunnelbroker is going to query for:
1. Token without instance assigned to it (new token or previous Tunnelbroker instance was gracefully shutdown)
2. Token where there wasn't heartbeat for some time interval, which means previous intance that owned token is dead.

For this purpose, we need two indexes on `assignedInstance` and `lastHeartbeat` to make it quicker and cheaper, we're going to query this a lot.

WE're using `userID` here, as this will be handy in Tunnelbroker, we need relationship between DCs token and `userID` to be able to find user devices and e.g. send notifs.

Depends on D15206

Test Plan: apply changes on terraform dev

Reviewers: bartek, tomek