[lib] Make sidebar config optional
Summary:
Farcaster threads can't have sidebars, so we should make the config optional. The problem with this approach is that it requires a couple of invariants. It's hard to avoid it completely, because even if we create a separate thread info, we still keep all the threads in the same Redux and SQLite collections.

Depends on D15209

Test Plan: Flow

Reviewers: bartek, kamil

Reviewed By: kamil

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D15213