[lib] Load Farcaster threads into the store
Summary:
When a user clicks a debug button, load and convert the threads, then add them into the store. At this point the threads are displayed, but opening them kills the app because most of the spec still isn't implemented.

Depends on D15214

Test Plan: Create a couple of threads in the Farcaster app. Click the debug button and check if the threads appear in the app with the correct names.

Reviewers: kamil, bartek

Differential Revision: https://phab.comm.dev/D15217