[native] Avoid keyboard in thread settings screen
Summary:
https://linear.app/comm/issue/ENG-11019/keyboardavoidingview-behavior

Moving KeyboardAvoidingView caused a regression on thread settings screen, the keyboard sometimes covers the screen.

We're using FlatList from react-native-gesture-handler in order to fix another bug (D6417) so just setting renderScrollComponent
won't work. So:
- we use FlatList from react-native
- which uses scroll view from react-native-keyboard-controller
- which uses scroll view from gesture-handler in order to fix the scrolling bug

Depends on D15239

Test Plan:
Click on description field in thread settings. The field should be scrolled up so that it's not covered by keyboard.
Make sure the screen is scrollable on Android.

Reviewers:

Subscribers: