[lib] Only insert the first portion of messages to the store
Summary:
The first batch of messages goes to both the DB and the store. The remaining ones go to the DB only.

Depends on D15251

https://linear.app/comm/issue/ENG-11172/fetch-all-the-farcaster-messages-after-connection

Test Plan: Fetch all the messages and check if only the first batch went to the store. Also modified the limit of initial messages to see if after opening the app all the messages are fetched from the DB.

Reviewers: bartek, kamil

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D15252