[lib] Resume fetching if the process was interrupted
Summary:
It is possible that a user kills the app before all the messages are fetched. This diff makes it so the process can be resumed.

Depends on D15256

https://linear.app/comm/issue/ENG-11172/fetch-all-the-farcaster-messages-after-connection

Test Plan: Connect and immediately kill the app. Open the app and check in the logs that the messages are being fetched. Open a thread with a lot of messages and check if they can be fetched from the DB.

Reviewers: bartek, kamil

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D15257