[lib] fix handling received messages twice
Summary:
Currently, after receiving `FARCASTER_CONNECTION_UPDATED` message, it is persisted (we do it for decrypted messages) and never removed, which causes this message to be processed on each app restart.

Additionally, added code to confirm it to the sender, so the sender can remove it from the Outbound P2P messages table.

Test Plan: Test the issues, it should be gone

Reviewers: tomek

Reviewed By: tomek

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D15261