[lib] fix handling received messages twice
Summary: Currently, after receiving `FARCASTER_CONNECTION_UPDATED` message is persisted (we do it for decrypted messages) and never remove, which causes this message to be processed on each app restart.

Additionally added code to confirm it to sender, so sender can remove it from Outbound P2P messages table.

Test Plan: Test the issues, it should be gone

Reviewers: tomek