[lib][web][native] handle update activities for Farcaster threads
Summary:
[ENG-11183](https://linear.app/comm/issue/ENG-11183/implement-farcasteractivityhandler)

Reusing the same logic as for DMs, with one change: for DMs, we want to send an update only when the thread was previously unread to avoid sending too many DM ops (which are encrypted and heavy). For Farcaster, I decided to send activity regardless of the previous state for two reasons:
1. Farcaster API is doing the same. Maybe it has a purpose on their backend, so I think it is safer to match the behaviour.
2. When opening the Farcaster thread, we fetch the thread info (D15258). This could cause a race condition when the response arrives after we set it to read, and can cause setting this as unread again, which is not great for UX. (Note that risk not exists for DM threads).

Made a note about this in [ENG-10729](https://linear.app/comm/issue/ENG-10729/consider-merging-activity-handlers).

Depends on D15267

Test Plan:
1. Test opening unread thread.
2. Test opening read thread.
3. Test marking the thread as unread when it is focused on a different device.
4. Test focuses thread when receiving a message.

Everything was confirmed with the Farcaster prod app.

Reviewers: tomek

Reviewed By: tomek

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D15268