[native] Show a modal asking about DCs connection
Summary:
https://linear.app/comm/issue/ENG-10993/show-a-modal-for-connected-farcaster-users-asking-about-dcs-connection

Modify the Farcaster connection modal to introduce a new step where users can connect their Farcaster account.

The fact that we can have users, who:
1. Didn't see the modal yet
2. Saw a modal in the past but dismissed it
3. Saw a modal and connected
4. Dismissed a modal and connected in the profile
And that we want to make the user experience good when they connect Farcaster and then DC using the modal, makes it really complicated to introduce the right logic to handle all of the options.

This diff is a reasonable tradeoff between functionality and complexity. The idea is that we reset the alert state for the Farcaster modal. This means that users who dismissed the modal in the past would need to do it again after they receive an update with the DC support - this could be both annoying and desirable, depending on the point of view.

The consequence of this diff is that:
1. Users who dismissed the modal in the past will see it again with an option to connect Farcaster first
2. Users who connected will see it again, but immediately on the DC step
3. Between the `supportsFarcasterDCs` flag is enabled and releasing this diff, users won't see the modal. It is a consequence of the migration we're doing - we clear the alert state, but we don't want to show the modal without DCs connection option.

Depends on D15093

Test Plan: Check if a modal is shown when an account isn't connected and connecting both levels is possible.
Check if a modal appears for existing Farcaster users who can now connect DCs.

Reviewers: kamil, bartek

Subscribers: ashoat

Differential Revision: https://phab.comm.dev/D15148