[lib] implement `useUsersSupportingProtocols` hook
Summary: [ENG-11204](https://linear.app/comm/issue/ENG-11204/add-thread-utils-to-decide-on-what-thread-to-create)

We need it to be able to check both thick and Farcaster thread support.

It is not great and can be improved but still, we need standalone functions in some places.

Depends on D15281

Test Plan: Tested later in the stack

Reviewers: tomek, bartek, angelika